Replace product card with NPIP test variation
