+++ client/components/jetpack/card/jetpack-product-card/style.scss
@@ -102,31 +102,65 @@
 }
 
 .jetpack-product-card__you-own-this {
-	margin-top: -18px;
-	margin-bottom: 31px;
-
-	padding-top: 6px;
-	padding-bottom: 6px;
-
+	margin-top: 40px;
+	margin-bottom: initial;
+	padding-top: 16px;
+	padding-bottom: 28px;
 	border-radius: ( 2px * 2 );
 
-	text-align: center;
-	background-color: rgba( var( --color-success-10-rgb ), 0.1 );
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+
+	color: var( --studio-gray-20 );
+	background-color: var( --studio-gray-0 );
+}
+
+.jetpack-product-card__you-own-this-icon {
+	display: block;
+
+	margin: 8px auto;
+	fill: currentColor;
 }
 
 .jetpack-product-card__price {
-	display: flex;
-	align-items: center;
-
-	margin: 16px 0 8px;
-
-	@include breakpoint-deprecated( '>660px' ) {
-		margin-top: -40px;
-	}
-
 	.is-disabled & {
 		opacity: 0.4;
 	}
+}
+
+.jetpack-product-card__price .plan-price.jetpack-product-card__original-price {
+	margin-inline-end: 16px;
+}
+
+.jetpack-product-card__you-save,
+.jetpack-product-card__get-started {
+	display: inline-block;
+	margin-block-start: 24px;
+	border-radius: 4px;
+	padding: 4px 24px 4px 8px;
+
+	font-size: $font-body-small;
+	font-weight: 700;
+
+	/* To make the slanted right edge */
+	clip-path: polygon(
+		0 0,
+		calc( 100% + 1px ) 0,
+		calc( 100% - 6px ) calc( 100% - 4px ),
+		calc( 100% - 8px ) calc( 100% - 1px ),
+		calc( 100% - 10px ) 100%,
+		0 100%
+	);
+}
+
+.jetpack-product-card__you-save {
+	background-color: var( --studio-yellow-10 );
+}
+
+.jetpack-product-card__get-started {
+	background-color: rgba( var( --studio-jetpack-green-10-rgb ), 0.2 );
 }
 
 .jetpack-product-card__price-from {
@@ -134,16 +168,33 @@
 	color: var( --studio-gray-40 );
 	font-size: 0.875rem;
 	line-height: 1.2;
+	vertical-align: text-top;
 }
 
 .jetpack-product-card .plan-price {
-	display: flex;
+	display: inline-flex;
 
 	color: var( --studio-gray-100 );
 }
 
-.jetpack-product-card .plan-price__currency-symbol {
-	margin-top: 0.875rem;
+.jetpack-product-card__price .plan-price.is-original {
+	color: var( --studio-gray-20 );
+
+	&::before {
+		border-top: 3px solid var( --studio-pink-50 );
+		border-radius: 3px;
+		transform: initial;
+	}
+}
+
+.jetpack-product-card__price .plan-price.is-discounted,
+.jetpack-product-card__price .plan-price.is-discounted .plan-price__currency-symbol {
+	color: inherit;
+}
+
+.jetpack-product-card__price .plan-price__currency-symbol {
+	margin-top: 0.25rem;
+	line-height: 3.375rem;
 
 	color: inherit;
 
@@ -157,13 +208,12 @@
 }
 
 .jetpack-product-card .plan-price__fraction {
-	margin-top: 1rem;
+	line-height: 3.375rem;
 }
 
 .jetpack-product-card__billing-time-frame,
 .jetpack-product-card__expiration-date {
-	margin: 0 12px;
-
+	display: block;
 	color: var( --studio-gray-40 );
 
 	font-size: 0.875rem;
@@ -171,6 +221,8 @@
 }
 
 .jetpack-product-card__price-tooltip {
+	margin-left: 4px;
+
 	color: var( --studio-gray-40 );
 	margin-bottom: 0.875rem;
 }
@@ -187,6 +239,10 @@
 }
 
 .jetpack-product-card__above-button {
+	position: relative;
+	top: 4px;
+	margin-bottom: calc( -0.875rem * 1.2 );
+
 	color: var( --studio-gray-40 );
 	font-size: 0.875rem;
 	line-height: 1.2;
@@ -202,9 +258,10 @@
 	height: 45px;
 
 	border-radius: ( 2px * 2 );
+	margin-block-start: 32px;
+	margin-block-end: 24px;
 
 	font-size: 1rem;
-	font-weight: 700;
 
 	em {
 		font-style: normal;
@@ -220,9 +277,7 @@
 }
 
 .jetpack-product-card__features-list {
-	margin: 32px 0;
-
-	list-style-type: none;
+	margin: 0 16px 32px;
 
 	em {
 		font-style: normal;
@@ -230,20 +285,9 @@
 }
 
 .jetpack-product-card__features-item {
-	display: flex;
-	align-items: flex-start;
-
 	margin: 14px 0;
-}
-
-.jetpack-product-card__features-icon {
-	margin-right: 12px;
-}
-
-.jetpack-product-card__features-text {
-	margin: 0;
-
-	color: var( --studio-gray-100 );
+	padding-inline-start: 8px;
+	color: var( --studio-gray-60 );
 }
 
 /* Prices Loading state */
@@ -261,5 +305,4 @@
 
 	width: 80px;
 	height: 16px;
-	margin-left: 12px;
 }
