--- client/components/jetpack/card/npip/jetpack-product-card-npip/features-item.tsx
@@ -1,24 +1,0 @@
-/**
- * External dependencies
- */
-import React from 'react';
-
-/**
- * Internal dependencies
- */
-import { preventWidows } from 'calypso/lib/formatting';
-
-/**
- * Type dependencies
- */
-import type { ProductCardFeaturesItem } from './types';
-
-interface Props {
-	item: ProductCardFeaturesItem;
-}
-
-const JetpackProductCardFeaturesItem: React.FC< Props > = ( { item: { text } } ) => (
-	<li className="jetpack-product-card-npip__features-item">{ preventWidows( text ) }</li>
-);
-
-export default JetpackProductCardFeaturesItem;
