--- client/components/jetpack/card/npip/jetpack-product-card-npip/style.scss
@@ -1,308 +1,0 @@
-.jetpack-product-card-npip {
-	position: relative;
-
-	box-sizing: border-box;
-	height: 100%;
-
-	background: var( --color-surface );
-	border: 1px solid var( --color-border-subtle );
-	border-radius: ( 2px * 4 );
-	border-radius: ( 2px * 4 );
-	box-shadow: 0 4px 24px rgba( 0, 0, 0, 0.05 );
-
-	&.is-featured.is-aligned {
-		position: relative;
-		z-index: 1;
-
-		margin-left: -8px;
-	}
-}
-
-.jetpack-product-card-npip__header {
-	display: flex;
-
-	box-sizing: border-box;
-	width: 100%;
-	padding: 8px;
-
-	background-color: var( --color-primary );
-	border-top-left-radius: inherit;
-	border-top-right-radius: inherit;
-	color: var( --color-text-inverted );
-
-	font-size: 0.875rem;
-	font-weight: 700;
-
-	.is-aligned & {
-		position: absolute;
-		bottom: 100%;
-		left: 0;
-	}
-}
-
-.jetpack-product-card-npip__header-icon {
-	margin-right: 8px;
-}
-
-.jetpack-product-card-npip__body {
-	padding: 24px 32px;
-
-	.is-featured & {
-		border: solid 2px var( --color-primary );
-		border-top: none;
-		border-bottom-left-radius: inherit;
-		border-bottom-right-radius: inherit;
-	}
-
-	.is-featured.is-aligned & {
-		box-sizing: border-box;
-		height: calc( 100% + 8px );
-
-		background-color: inherit;
-	}
-}
-
-.jetpack-product-card-npip__icon {
-	width: 56px;
-}
-
-.jetpack-product-card-npip__product-name {
-	margin: 6px 0 12px;
-
-	color: var( --studio-gray-100 );
-
-	font-family: 'Noto Sans SemiCondensed Bold', 'Noto Sans', $sans;
-	font-size: 2rem;
-	font-weight: 700;
-	line-height: 1.2;
-
-	.without-icon & {
-		margin: 12px 0 20px;
-	}
-
-	em {
-		font-style: normal;
-	}
-
-	.is-disabled & {
-		opacity: 0.4;
-	}
-}
-
-.jetpack-product-card-npip__description {
-	color: var( --studio-gray-60 );
-
-	@include breakpoint-deprecated( '>660px' ) {
-		min-height: 108px;
-	}
-
-	.is-disabled & {
-		opacity: 0.4;
-	}
-}
-
-.jetpack-product-card-npip__you-own-this {
-	margin-top: 40px;
-	margin-bottom: initial;
-	padding-top: 16px;
-	padding-bottom: 28px;
-	border-radius: ( 2px * 2 );
-
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-
-	color: var( --studio-gray-20 );
-	background-color: var( --studio-gray-0 );
-}
-
-.jetpack-product-card-npip__you-own-this-icon {
-	display: block;
-
-	margin: 8px auto;
-	fill: currentColor;
-}
-
-.jetpack-product-card-npip__price {
-	.is-disabled & {
-		opacity: 0.4;
-	}
-}
-
-.jetpack-product-card-npip__price .plan-price.jetpack-product-card-npip__original-price {
-	margin-inline-end: 16px;
-}
-
-.jetpack-product-card-npip__you-save,
-.jetpack-product-card-npip__get-started {
-	display: inline-block;
-	margin-block-start: 24px;
-	border-radius: 4px;
-	padding: 4px 24px 4px 8px;
-
-	font-size: $font-body-small;
-	font-weight: 700;
-
-	/* To make the slanted right edge */
-	clip-path: polygon(
-		0 0,
-		calc( 100% + 1px ) 0,
-		calc( 100% - 6px ) calc( 100% - 4px ),
-		calc( 100% - 8px ) calc( 100% - 1px ),
-		calc( 100% - 10px ) 100%,
-		0 100%
-	);
-}
-
-.jetpack-product-card-npip__you-save {
-	background-color: var( --studio-yellow-10 );
-}
-
-.jetpack-product-card-npip__get-started {
-	background-color: rgba( var( --studio-jetpack-green-10-rgb ), 0.2 );
-}
-
-.jetpack-product-card-npip__price-from {
-	margin-right: 12px;
-	color: var( --studio-gray-40 );
-	font-size: 0.875rem;
-	line-height: 1.2;
-	vertical-align: text-top;
-}
-
-.jetpack-product-card-npip .plan-price {
-	display: inline-flex;
-
-	color: var( --studio-gray-100 );
-}
-
-.jetpack-product-card-npip__price .plan-price.is-original {
-	color: var( --studio-gray-20 );
-
-	&::before {
-		border-top: 3px solid var( --studio-pink-50 );
-		border-radius: 3px;
-		transform: initial;
-	}
-}
-
-.jetpack-product-card-npip__price .plan-price.is-discounted,
-.jetpack-product-card-npip__price .plan-price.is-discounted .plan-price__currency-symbol {
-	color: inherit;
-}
-
-.jetpack-product-card-npip__price .plan-price__currency-symbol {
-	margin-top: 0.25rem;
-	line-height: 3.375rem;
-
-	color: inherit;
-
-	font-size: 1.5rem;
-}
-
-.jetpack-product-card-npip__price-free,
-.jetpack-product-card-npip .plan-price__integer {
-	font-size: 3.375rem;
-	font-weight: 700;
-}
-
-.jetpack-product-card-npip .plan-price__fraction {
-	line-height: 3.375rem;
-}
-
-.jetpack-product-card-npip__billing-time-frame,
-.jetpack-product-card-npip__expiration-date {
-	display: block;
-	color: var( --studio-gray-40 );
-
-	font-size: 0.875rem;
-	line-height: 1.2;
-}
-
-.jetpack-product-card-npip__price-tooltip {
-	margin-left: 4px;
-
-	color: var( --studio-gray-40 );
-	margin-bottom: 0.875rem;
-}
-
-.jetpack-product-card-npip__price-tooltip.popover .popover__inner {
-	text-align: center;
-}
-.jetpack-product-card-npip__price-tooltip p {
-	margin-bottom: 1rem;
-}
-
-.jetpack-product-card-npip__expiration-date {
-	color: var( --studio-pink-50 );
-}
-
-.jetpack-product-card-npip__above-button {
-	position: relative;
-	top: 4px;
-	margin-bottom: calc( -0.875rem * 1.2 );
-
-	color: var( --studio-gray-40 );
-	font-size: 0.875rem;
-	line-height: 1.2;
-}
-
-.jetpack-product-card-npip__disabled-message {
-	color: var( --color-accent );
-	font-weight: 600;
-}
-
-.jetpack-product-card-npip__button {
-	width: 100%;
-	height: 45px;
-
-	border-radius: ( 2px * 2 );
-	margin-block-start: 32px;
-	margin-block-end: 24px;
-
-	font-size: 1rem;
-
-	em {
-		font-style: normal;
-	}
-}
-
-.jetpack-product-card-npip__features {
-	margin-bottom: 10px;
-
-	.is-disabled & {
-		opacity: 0.4;
-	}
-}
-
-.jetpack-product-card-npip__features-list {
-	margin: 0 16px 32px;
-
-	em {
-		font-style: normal;
-	}
-}
-
-.jetpack-product-card-npip__features-item {
-	margin: 14px 0;
-	padding-inline-start: 8px;
-	color: var( --studio-gray-60 );
-}
-
-/* Prices Loading state */
-.jetpack-product-card-npip__price-placeholder {
-	@include placeholder( --color-neutral-10 );
-
-	width: 80px;
-	height: 40px;
-	margin-bottom: 16px;
-	margin-top: 24px;
-}
-
-.jetpack-product-card-npip__time-frame-placeholder {
-	@include placeholder( --color-neutral-10 );
-
-	width: 80px;
-	height: 16px;
-}
