--- client/components/jetpack/card/npip/jetpack-product-card-npip/types.ts
@@ -1,14 +1,0 @@
-/**
- * Internal dependencies
- */
-
-/**
- * Type dependencies
- */
-import type {
-	SelectorProductFeaturesItem,
-	SelectorProductFeatures,
-} from 'calypso/my-sites/plans/jetpack-plans/types';
-
-export type ProductCardFeaturesItem = SelectorProductFeaturesItem;
-export type ProductCardFeatures = SelectorProductFeatures;
