+++ client/jetpack-cloud/sections/pricing/style.scss
@@ -67,15 +67,12 @@
 
 	// On cloud.jetpack.com, we want to override the primary color, used on
 	// the featured product card, with --studio-pink-50.
-	.jetpack-product-card.is-featured,
-	.jetpack-product-card-npip.is-featured {
-		.jetpack-product-card__header,
-		.jetpack-product-card-npip__header {
+	.jetpack-product-card.is-featured {
+		.jetpack-product-card__header {
 			background-color: var( --studio-pink-50 );
 		}
 
-		.jetpack-product-card__body,
-		.jetpack-product-card-npip__body {
+		.jetpack-product-card__body {
 			border: solid 2px var( --studio-pink-50 );
 		}
 	}
